Improved default setting of docker_hostname
If the LK_DOCKER_HOST environment variable is not set, try to get the hostname from the os.Hostname() function.
If that fails, only then set the hostname to 'unknown-docker-host'.  This is useful for loghost when deployed in
swarm mode, since you won't know what hostname it's running on when you start the service.